server/requirements.txt
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
drf-yasg==1.20.0
psycopg2-binary>=2.9.1,<2.9.2
drf-yasg==1.20.0
djangorestframework-simplejwt==5.0.0
djoser==2.0.5
django-jazzmin==2.4.8
gunicorn==20.1.0
whitenoise==5.3.0
JustWatch==0.5.1
Pillow==8.3.1